requirements.txt
amqp==5.1.1
asgiref==3.7.2
autopep8==2.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-rest-framework==0.1.0
django-silk==5.0.3
djangorestframework==3.14.0
google-play-scraper==1.2.4
gprof2dot==2022.7.29
idna==3.4
kombu==5.3.1
numpy==1.25.0
pandas==2.0.2
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
